Split moves.py into per-piece package for easier AI agent and developer readability
